+++ test/document.test.js
@@ -8725,4 +8725,77 @@ describe('document', function() {
       assert.equal(fromDb.placedItems['1'].image, 'updated');
     });
   });
+
+  it('setting nested array path to non-nested array wraps values top-down (gh-8544)', function() {
+    const positionSchema = mongoose.Schema({
+      coordinates: {
+        type: [[Number]],
+        required: true
+      },
+      lines: {
+        type: [[[Number]]],
+        required: true
+      }
+    });
+
+    const Position = db.model('Test', positionSchema);
+    const position = new Position();
+
+    position.coordinates = [1, 2];
+    position.lines = [3, 4];
+
+    const obj = position.toObject();
+    assert.deepEqual(obj.coordinates, [[1, 2]]);
+    assert.deepEqual(obj.lines, [[[3, 4]]]);
+  });
+
+  it('doesnt wipe out nested keys when setting nested key to empty object with minimize (gh-8565)', function() {
+    const opts = { autoIndex: false, autoCreate: false };
+    const schema1 = Schema({ plaid: { nestedKey: String } }, opts);
+    const schema2 = Schema({ plaid: { nestedKey: String } }, opts);
+    const schema3 = Schema({ plaid: { nestedKey: String } }, opts);
+
+    const Test1 = db.model('Test1', schema1);
+    const Test2 = db.model('Test2', schema2);
+    const Test3 = db.model('Test3', schema3);
+
+    const doc1 = new Test1({});
+    assert.deepEqual(doc1.toObject({ minimize: false }).plaid, {});
+
+    const doc2 = new Test2({ plaid: doc1.plaid });
+    assert.deepEqual(doc2.toObject({ minimize: false }).plaid, {});
+
+    const doc3 = new Test3({});
+    doc3.set({ plaid: doc2.plaid });
+    assert.deepEqual(doc3.toObject({ minimize: false }).plaid, {});
+  });
+
+  it('allows calling `validate()` in post validate hook without causing parallel validation error (gh-8597)', function() {
+    const EmployeeSchema = Schema({
+      name: String,
+      employeeNumber: {
+        type: String,
+        validate: v => v.length > 5
+      }
+    });
+    let called = 0;
+
+    EmployeeSchema.post('validate', function() {
+      ++called;
+      if (!this.employeeNumber && !this._employeeNumberRetrieved) {
+        this.employeeNumber = '123456';
+        this._employeeNumberRetrieved = true;
+        return this.validate();
+      }
+    });
+
+    const Employee = db.model('Test', EmployeeSchema);
+
+    return co(function*() {
+      const e = yield Employee.create({ name: 'foo' });
+      assert.equal(e.employeeNumber, '123456');
+      assert.ok(e._employeeNumberRetrieved);
+      assert.equal(called, 2);
+    });
+  });
 });